requirements.txt
altair==5.5.0
ipyleaflet>=0.14.0
ipywidgets>=7.7.1
matplotlib>=3.7.0
numpy>=1.24.2
opencv-python>=4.1.2
pandas==3.0.0
pyarrow==23.0.1
psycopg2-binary>=2.9.3
pymysql>=1.0.2
python-dotenv>=0.20.0
requests>=2.27.1
scikit-learn
seaborn>=0.12.2
sqlalchemy>=2.0.38
streamlit==1.54.0
sympy>=1.10.1
xgboost
deep-translator==1.11.4
langdetect==1.0.9
sentence_transformers==5.2.3
faiss-cpu==1.13.2
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.2+cu121
